refactor: AcceptanceTest 불필요 메서드 각 Test Class 로 이동
