+++ src/test/java/wooteco/subway/admin/acceptance/AcceptanceTest.java
@@ -8,12 +8,7 @@ import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
-import wooteco.subway.admin.dto.LineResponse;
-import wooteco.subway.admin.dto.StationResponse;
 
-import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -62,41 +57,6 @@ public class AcceptanceTest {
     public static RequestSpecification given() {
         return RestAssured.given().log().all();
     }
-
-    LineResponse createLine(String name) {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", name);
-        params.put("startTime", LocalTime.of(5, 30).format(DateTimeFormatter.ISO_LOCAL_TIME));
-        params.put("endTime", LocalTime.of(23, 30).format(DateTimeFormatter.ISO_LOCAL_TIME));
-        params.put("intervalTime", "10");
-        String path = "lines";
-        return post(path, params, LineResponse.class);
-
-    }
-
-    StationResponse createStation(String name) {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", name);
-        String path = "/stations";
-        return post(path, params, StationResponse.class);
-    }
-
-    void addLineStation(Long lineId, Long preStationId, Long stationId) {
-        addLineStation(lineId, preStationId, stationId, 10, 10);
-    }
-
-    Void addLineStation(Long lineId, Long preStationId, Long stationId, Integer distance, Integer duration) {
-        Map<String, String> params = new HashMap<>();
-        params.put("preStationId", preStationId == null ? "" : preStationId.toString());
-        params.put("stationId", stationId.toString());
-        params.put("distance", distance.toString());
-        params.put("duration", duration.toString());
-        String path = "/lines/" + lineId + "/stations";
-        return post(path, params, Void.class);
-
-
-    }
-
     <T> T post(String path, Map<String, String> params, Class<T> responseType) {
         if (responseType.equals(Void.class)) {
             okStatusPost(path, params);
