+++ src/test/java/wooteco/subway/admin/acceptance/LineStationAcceptanceTest.java
@@ -7,6 +7,11 @@ import wooteco.subway.admin.dto.LineDetailResponse;
 import wooteco.subway.admin.dto.LineResponse;
 import wooteco.subway.admin.dto.StationResponse;
 
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+import java.util.HashMap;
+import java.util.Map;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 @Sql({"/schema-test.sql","/truncate.sql"})
@@ -34,6 +39,23 @@ public class LineStationAcceptanceTest extends AcceptanceTest {
         assertThat(lineResponseAfterRemoveLineStation.getStations().size()).isEqualTo(2);
     }
 
+    LineResponse createLine(String name) {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", name);
+        params.put("startTime", LocalTime.of(5, 30).format(DateTimeFormatter.ISO_LOCAL_TIME));
+        params.put("endTime", LocalTime.of(23, 30).format(DateTimeFormatter.ISO_LOCAL_TIME));
+        params.put("intervalTime", "10");
+        String path = "lines";
+        return post(path, params, LineResponse.class);
+    }
+
+    StationResponse createStation(String name) {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", name);
+        String path = "/stations";
+        return post(path, params, StationResponse.class);
+    }
+
     LineDetailResponse getLine(Long id) {
         String path = "/lines/" + id;
         return get(path, LineDetailResponse.class);
@@ -43,4 +65,19 @@ public class LineStationAcceptanceTest extends AcceptanceTest {
         String path = "/lines/" + lineId + "/stations/" + stationId;
         delete(path);
     }
+
+    Void addLineStation(Long lineId, Long preStationId, Long stationId, Integer distance, Integer duration) {
+        Map<String, String> params = new HashMap<>();
+        params.put("preStationId", preStationId == null ? "" : preStationId.toString());
+        params.put("stationId", stationId.toString());
+        params.put("distance", distance.toString());
+        params.put("duration", duration.toString());
+        String path = "/lines/" + lineId + "/stations";
+        return post(path, params, Void.class);
+    }
+
+    void addLineStation(Long lineId, Long preStationId, Long stationId) {
+        addLineStation(lineId, preStationId, stationId, 10, 10);
+    }
+
 }
