+++ src/test/java/wooteco/subway/admin/acceptance/StationAcceptanceTest.java
@@ -5,7 +5,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.test.context.jdbc.Sql;
 import wooteco.subway.admin.dto.StationResponse;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -27,6 +29,13 @@ public class StationAcceptanceTest extends AcceptanceTest {
         // then
         List<StationResponse> stationsAfterDelete = getStations();
         assertThat(stationsAfterDelete.size()).isEqualTo(2);
+    }
+
+    StationResponse createStation(String name) {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", name);
+        String path = "/stations";
+        return post(path, params, StationResponse.class);
     }
 
     List<StationResponse> getStations() {
